Update follows controller destroy method to not delete follow if follow is user.  Add @self variable to posts show method, use variable to hide follow/unfollow links on users own posts.

+++ app/controllers/follows_controller.rb
@@ -18,8 +18,8 @@ class FollowsController < ApplicationController
   
   def destroy
     follow = Follow.find_by(followed_user_id: params[:id])
-    p "=====================#{follow.followed_user_id}"
-    follow.destroy
+
+    follow.destroy if params[:id] != current_user.id
     redirect_to list_follow_posts_path
   end
 end
